+++ app/src/main/java/com/mmdev/roove/ui/MainActivity.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 27.02.20 15:42
+ * Last modified 27.02.20 15:51
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -28,7 +28,6 @@ import com.mmdev.roove.ui.core.SharedViewModel
 import com.mmdev.roove.ui.core.viewmodel.RemoteUserRepoViewModel
 import com.mmdev.roove.ui.custom.LoadingDialog
 import com.mmdev.roove.ui.main.MainFlowFragment
-import com.mmdev.roove.utils.UtilityManager
 import com.mmdev.roove.utils.doOnApplyWindowInsets
 import com.mmdev.roove.utils.observeOnce
 import kotlinx.android.synthetic.main.activity_main.*
@@ -108,10 +107,10 @@ class MainActivity: AppCompatActivity(R.layout.activity_main) {
 
 		//creating fake users, do not call this on UI thread
 		//UtilityManager.createFakeUsersOnRemote()
-		for (i in 0 until 20){
-			UtilityManager.generateConversationOnRemote()
-			UtilityManager.generateMatchesOnRemote()
-		}
+//		for (i in 0 until 20){
+//			UtilityManager.generateConversationOnRemote()
+//			UtilityManager.generateMatchesOnRemote()
+//		}
 
 
 	}
