+++ app/src/main/java/com/mmdev/roove/ui/auth/view/RegistrationFragment.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 26.02.20 17:22
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -21,8 +21,8 @@ import androidx.constraintlayout.motion.widget.MotionLayout
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
-import com.mmdev.business.user.BaseUserInfo
-import com.mmdev.business.user.UserItem
+import com.mmdev.business.core.BaseUserInfo
+import com.mmdev.business.core.UserItem
 import com.mmdev.roove.R
 import com.mmdev.roove.ui.auth.AuthViewModel
 import com.mmdev.roove.ui.core.BaseFragment
@@ -214,15 +214,16 @@ class RegistrationFragment: BaseFragment(R.layout.fragment_registration){
 
 		btnRegistrationDone.setOnClickListener {
 			val finalUserModel = BaseUserInfo(name,
-			                                  age,
-			                                  city,
-			                                  gender,
-			                                  preferredGender,
-			                                  baseUserInfo.mainPhotoUrl,
-			                                  baseUserInfo.userId)
+			                                                                             age,
+			                                                                             city,
+			                                                                             gender,
+			                                                                             preferredGender,
+			                                                                             baseUserInfo.mainPhotoUrl,
+			                                                                             baseUserInfo.userId)
 
 			authViewModel.signUp(UserItem(finalUserModel,
-			                              photoURLs = mutableListOf(finalUserModel.mainPhotoUrl)))
+			                                                                         photoURLs = mutableListOf(
+					                                                                         finalUserModel.mainPhotoUrl)))
 		}
