+++ app/src/main/res/layout/fragment_cards_item.xml
@@ -2,7 +2,7 @@
 <!--
   ~ Created by Andrii Kovalchuk
   ~ Copyright (c) 2020. All rights reserved.
-  ~ Last modified 16.02.20 17:30
+  ~ Last modified 27.02.20 15:53
   ~
   ~ This Source Code Form is subject to the terms of the Mozilla Public
   ~ License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -17,7 +17,7 @@
 
 		<variable
 			name="userItem"
-			type="com.mmdev.business.user.UserItem" />
+			type="com.mmdev.business.core.UserItem" />
 
 	</data>
