+++ business/src/main/java/com/mmdev/business/auth/AuthUserItem.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 16.02.20 16:33
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,7 +10,7 @@
 
 package com.mmdev.business.auth
 
-import com.mmdev.business.user.BaseUserInfo
+import com.mmdev.business.core.BaseUserInfo
 
 /**
  * This is the documentation block about the class
