+++ business/src/main/java/com/mmdev/business/auth/repository/AuthRepository.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 16.02.20 16:05
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,7 +10,8 @@
 
 package com.mmdev.business.auth.repository
 
-import com.mmdev.business.user.UserItem
+import com.mmdev.business.core.BaseUserInfo
+import com.mmdev.business.core.UserItem
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Single
@@ -20,7 +21,7 @@ interface AuthRepository {
 
 	fun isAuthenticatedListener(): Observable<Boolean>
 
-	fun signIn(token: String): Single<HashMap<Boolean, UserItem>>
+	fun signIn(token: String): Single<HashMap<Boolean, BaseUserInfo>>
 
 	fun registerUser(userItem: UserItem): Completable
