+++ business/src/main/java/com/mmdev/business/cards/usecase/AddToSkippedUseCase.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 16.02.20 17:25
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -11,7 +11,7 @@
 package com.mmdev.business.cards.usecase
 
 import com.mmdev.business.cards.repository.CardsRepository
-import com.mmdev.business.user.UserItem
+import com.mmdev.business.core.UserItem
 
 /**
  * This is the documentation block about the class
