+++ business/src/main/java/com/mmdev/business/chat/entity/MessageItem.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 02.02.20 20:08
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,7 +10,7 @@
 
 package com.mmdev.business.chat.entity
 
-import com.mmdev.business.user.BaseUserInfo
+import com.mmdev.business.core.BaseUserInfo
 
 data class MessageItem(val sender: BaseUserInfo = BaseUserInfo(),
                        val recipientId: String = "",
