+++ business/src/main/java/com/mmdev/business/conversations/ConversationItem.kt
@@ -1,7 +1,7 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 17.02.20 14:57
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,7 +10,7 @@
 
 package com.mmdev.business.conversations
 
-import com.mmdev.business.user.UserItem
+import com.mmdev.business.core.UserItem
 import java.util.*
 
 /**
