+++ business/src/main/java/com/mmdev/business/remote/usecase/UpdateUserItemUseCase.kt
@@ -1,24 +1,24 @@
 /*
  * Created by Andrii Kovalchuk
  * Copyright (c) 2020. All rights reserved.
- * Last modified 27.02.20 15:31
+ * Last modified 27.02.20 15:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package com.mmdev.business.user.usecase.remote
+package com.mmdev.business.remote.usecase
 
-import com.mmdev.business.user.BaseUserInfo
-import com.mmdev.business.user.repository.RemoteUserRepository
+import com.mmdev.business.core.UserItem
+import com.mmdev.business.remote.RemoteUserRepository
 
 /**
  * This is the documentation block about the class
  */
 
-class GetFullUserInfoUseCase (private val repository: RemoteUserRepository) {
+class UpdateUserItemUseCase(private val repository: RemoteUserRepository) {
 
-	fun execute(t: BaseUserInfo) = repository.getFullUserItem(t)
+	fun execute(u: UserItem) = repository.updateUserItem(u)
 
 }
